Ensure uploaded files keep their names
When migrating to dexador, we lost the ability of explicitly setting the
filename when uploading a file via POST request. We work around this by
creating a symlink, with the right filename, before uploading.

Setting upload filenames properly allows GitLab's auto image display do
do the right thing, for instance.