Don't require all users to be admins to migrate objects
Considering the length of the migration process, escalating all users as
admins to run some API commands is too intrusive. As we're already doing
a lot of operations in the rails console for every mapped object (to
update the activity timeline and other ancillary objects), we might as
well update the object dates and authors directly as well. Doing so
allows us not to give any users admin status in the course of the migration.